+++ flask_nacos/__init__.py
@@ -11,7 +11,7 @@ from .exceptions import (
 )
 from .extension import FlaskNacos
 
-__version__ = "1.0.0"
+__version__ = "1.0.1"
 
 __all__ = [
     "FlaskNacos",